[#32] 패키지 계층 리팩토링
