+++ src/main/java/zeobase/zbtechnical/challenges/controller/ReservationController.java
@@ -8,9 +8,9 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 import zeobase.zbtechnical.challenges.dto.reservation.ReservationAccept;
-import zeobase.zbtechnical.challenges.dto.reservation.ReservationAvailableResponse;
+import zeobase.zbtechnical.challenges.dto.reservation.response.ReservationAvailableResponse;
 import zeobase.zbtechnical.challenges.dto.reservation.ReservationReserve;
-import zeobase.zbtechnical.challenges.dto.reservation.ReservationInfoResponse;
+import zeobase.zbtechnical.challenges.dto.reservation.response.ReservationInfoResponse;
 import zeobase.zbtechnical.challenges.exception.StoreException;
 import zeobase.zbtechnical.challenges.service.ReservationService;
 import zeobase.zbtechnical.challenges.type.ErrorCode;
