+++ src/main/java/zeobase/zbtechnical/challenges/dto/member/request/MemberSigninRequest.java
@@ -1,10 +1,10 @@
-package zeobase.zbtechnical.challenges.dto.member;
+package zeobase.zbtechnical.challenges.dto.member.request;
 
 import lombok.Builder;
 import lombok.Getter;
 
 /**
- * 이용자의 로그인 관련 DTO 클래스
+ * 이용자의 로그인 관련 request DTO 클래스
  */
 @Getter
 @Builder
