+++ src/main/java/zeobase/zbtechnical/challenges/dto/reservation/response/ReservationAvailableResponse.java
@@ -1,4 +1,4 @@
-package zeobase.zbtechnical.challenges.dto.reservation;
+package zeobase.zbtechnical.challenges.dto.reservation.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,7 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 예약 가능 여부 관련 DTO 클래스
+ * 예약 가능 여부 관련 response DTO 클래스
  */
 @Getter
 @Builder
