+++ src/main/java/zeobase/zbtechnical/challenges/dto/review/response/ReviewInfoResponse.java
@@ -1,4 +1,4 @@
-package zeobase.zbtechnical.challenges.dto.review;
+package zeobase.zbtechnical.challenges.dto.review.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
