+++ src/main/java/zeobase/zbtechnical/challenges/service/StoreService.java
@@ -1,8 +1,8 @@
 package zeobase.zbtechnical.challenges.service;
 
 import org.springframework.security.core.Authentication;
-import zeobase.zbtechnical.challenges.dto.store.StoreDistanceInfoResponse;
-import zeobase.zbtechnical.challenges.dto.store.StoreInfoResponse;
+import zeobase.zbtechnical.challenges.dto.store.response.StoreDistanceInfoResponse;
+import zeobase.zbtechnical.challenges.dto.store.response.StoreInfoResponse;
 import zeobase.zbtechnical.challenges.dto.store.StoreRegistration;
 
 import java.util.List;
