+++ src/main/java/zeobase/zbtechnical/challenges/service/impl/ReservationServiceImpl.java
@@ -6,9 +6,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import zeobase.zbtechnical.challenges.dto.reservation.ReservationAccept;
-import zeobase.zbtechnical.challenges.dto.reservation.ReservationAvailableResponse;
+import zeobase.zbtechnical.challenges.dto.reservation.response.ReservationAvailableResponse;
 import zeobase.zbtechnical.challenges.dto.reservation.ReservationReserve;
-import zeobase.zbtechnical.challenges.dto.reservation.ReservationInfoResponse;
+import zeobase.zbtechnical.challenges.dto.reservation.response.ReservationInfoResponse;
 import zeobase.zbtechnical.challenges.entity.Member;
 import zeobase.zbtechnical.challenges.entity.Reservation;
 import zeobase.zbtechnical.challenges.entity.Store;
