+++ src/main/java/zeobase/zbtechnical/challenges/service/impl/StoreServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import zeobase.zbtechnical.challenges.dto.store.StoreDistanceInfoResponse;
-import zeobase.zbtechnical.challenges.dto.store.StoreInfoResponse;
+import zeobase.zbtechnical.challenges.dto.store.response.StoreDistanceInfoResponse;
+import zeobase.zbtechnical.challenges.dto.store.response.StoreInfoResponse;
 import zeobase.zbtechnical.challenges.dto.store.StoreRegistration;
 import zeobase.zbtechnical.challenges.entity.Member;
 import zeobase.zbtechnical.challenges.entity.Store;
